swaynag: split config into own file and fix optind
